backend/requirements.txt
pytest~=6.2.2
pytest-mock~=3.5.1
SQLAlchemy~=1.3.23
requests~=2.25.1
humps~=0.2.2
fastapi-mail~=0.3.4.2
pandas~=1.2.2